requirements.txt
Django>=5.2.1
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
django-extensions>=3.2.3

# Production dependencies
dj-database-url>=2.1.0
psycopg2-binary>=2.9.7
redis>=4.6.0
gunicorn>=21.2.0
whitenoise>=6.5.0